requirements.txt
# 🚀 Professional Trading Bot - Dependencias
# ==========================================

# Core Framework
tensorflow-macos==2.15.0
tensorflow-metal
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Trading & Market Data
python-binance>=1.0.19
aiohttp>=3.8.0
# asyncio-mqtt>=0.13.0  # Opcional para MQTT connections

# Technical Analysis
ta>=0.10.2
# talib-binary>=0.4.26  # Opcional - requiere compilación manual en algunos sistemas

# Database & Storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Data Processing
pydantic>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Monitoring
structlog>=23.0.0
rich>=13.0.0
colorama>=0.4.6

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0

# HTTP & API
requests>=2.31.0
httpx>=0.24.0

# Math & Statistics
scipy>=1.11.0
statsmodels>=0.14.0

# Async & Concurrency
aiofiles>=23.0.0
asyncio>=3.4.3

# Development & Testing (opcional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Discord Integration
discord.py>=2.3.0
discord-webhook>=1.3.0

# Risk Management (opcional)
# python-telegram-bot>=20.0  # Solo si usas notificaciones Telegram

# Utilities
click>=8.1.0
tqdm>=4.65.0
ujson>=5.8.0

# Apple Silicon Optimization (para Mac M1/M2/M3)
# Estas se instalan automáticamente en macOS con Apple Silicon
